Move common utility files to ascent.utils
Update new import paths.
