+++ ascent/datamodules/nnunet_reg_datamodule.py
@@ -13,13 +13,13 @@ from monai.transforms import (
     SpatialPadd,
 )
 
-from ascent.datamodules.components.transforms import (
+from ascent.datamodules.nnunet_datamodule import nnUNetDataModule
+from ascent.utils.transforms import (
     Convert2Dto3Dd,
     Convert3Dto2Dd,
     LoadNpyd,
     MayBeSqueezed,
 )
-from ascent.datamodules.nnunet_datamodule import nnUNetDataModule
 
 
 class nnUNetRegDataModule(nnUNetDataModule):
