+++ configs/model/unwrap_2d.yaml
@@ -10,14 +10,14 @@ optimizer:
   weight_decay: 3e-05
 
 scheduler:
-  _target_: ascent.models.components.unet_related.poly_lr.PolynomialLR
+  _target_: ascent.utils.scheduler.poly_lr.PolynomialLR
   _partial_: true
   max_decay_steps: ${trainer.max_epochs}
 
 net:
   _target_: ascent.models.components.spyrit_net.SpyritNet
   unwrap:
-    _target_: ascent.models.components.spyrit_related.utils.Robust2DUnwrap
+    _target_: ascent.models.components.spyrit_related.unwrap.Robust2DUnwrap
     shape: [40, 192]
     mu: 1e-6
